Improve transaction merging
Transaction merging does not provide any visual feedback to the user.
This can lead to unwanted results. This mostly happeqns, when the user
selected transactions appear in an unexpected order to the matching
function and the category assignment disappears for no obvious reason.

The logic has been enhanced to check if an imported transaction which
usually does not have a category assigned shall replace an existing
transaction that has a category assignment. In this case, the selected
transactions are exchanged so that the category assignment does not
disappear. In case both transactions have more than one split, the user
is asked for confirmation of the operation.

BUG: 435512
FIXED-IN: 5.1.2

(cherry picked from commit 49e4e13a546a5b6bdfbbbcef65bdb59cc54ae4ab)