Solve problems with usage of QTemporaryFile on MS-Windows
A call to QTemporaryFile::close() may not close the file but only rewind
the file pointer to the beginning allwing faster re-open. This causes
problems on MS-Windows filesystems, as one cannot e.g. rename the
temporary file in this case because it is still kept open from the
perspective of the filesystem. Only destroying the object really closes
the file on the filesystem.

This change makes sure, that the QTemporaryFile object only lives as
long as it is needed and closed immediately after by destroying the
object.

BUG: 436647
FIXED-IN: 5.1.2
(cherry picked from commit 39785a3d80c56dc6e53f301d76c41bcc8ca048c6)