Added info icons and tooltip to ledger view
Icons are now shown on the right side of the details column for
erroneous, imported and matched transactions as well as overdue
scheduled transactions. Hovering over the icon shows additional
information. Clicking on the icon, except on an erroneous transaction
provides access to an action dialog that allows to accept or unmatch the
transaction or enter the schedule.