Show text under icons in toolbar, not next to them
The default behavior is to show the text next to the toolbar icons.
However, since we enable quite some of them by default, the
toolbar is too small to show all of them. As a result, those that
do _not_ fit can only be accessed by pressing the "≫" extension
button.

This is suboptimal and can be solved by switch to showing the text
_under_ the icon, allowing all of the icons to be visible.