Improve commit 4fcb0ab18c2bc660f9a781bafcd8e09ca96f638e
The change in strategy causes the file to be saved improperly if
attribute data contains characters that are not valid according to the
XML specifications. We received reports that such data can exist due to
import of transactions from external sources.

https://invent.kde.org/office/kmymoney/-/issues/45 and
https://mail.kde.org/pipermail/kmymoney-devel/2021-June/027387.html
contain the reports.

This change removes all such characters before they are used as
attribute values.