Preset parent account with the current selected account
When having selected an account in the accounts view and one selects to
create a new one it should as a child account to the current one. The
hierarchy page however did not get set up properly so that the user had
to select the account again.

This change fixes the behavior so that the parent is selected in the
hierarchy view of the new account wizard.