UX/UI: Move "Find" transaction button to the right
"Find" button in Find Transaction Dialog is a positive action which ends
 up in some action getting executed (AcceptRole). Most OS (if not all)
would have it placed to the right in the dialog. Using "Apply" button
(ApplyRole) for its purpose does, instead, place on the left, which goes
 against typical user's intuition.

Fix this by using "Open" button instead "Apply". The implied AcceptRole
implicitly makes it a default button, too, so no need for the
setDefault() redundancy in the code.