Make sure to load all templates into resource file
Only templates in directories with a language but no country were
included on Linux. Quote from
https://cmake.org/cmake/help/latest/command/file.html#glob-recurse

On Windows and macOS, globbing is case-insensitive even if the
underlying filesystem is case-sensitive (both filenames and globbing
expressions are converted to lowercase before matching). On other
platforms, globbing is case-sensitive.

This change will include the missing templates on Linux as well.