Don't remove payee if still referenced
The statement importer creates payees on the fly if they are not known
and destroys all those after the statement processing is done that are
not referenced. Since the reference check was not performed all of them
were removed at the end of the import.

This change forces to throw an exception when a payee is still
referenced when it should be removed which will take care of keeping it
after the import of the statement.

BUG: 441696