Improve search for online task plugins
By putting them in their own namespace we don't need to list/check
all available plugins from the "kmymoney" namespace.

This will also help getting rid of the service types, because this was the only
plugin in the "kmymoney" namespace which did not contain the "KMyMoney/Plugin"
service type.

Relates to https://phabricator.kde.org/T14483