Do not embed unneeded json metadata in KCMs
This will slow up the build time & blot up the binaries.

When we can depend on the latest KCoreAddons (which might take a while) we can utilize
the K_PLUGIN_CLASS macro (https://invent.kde.org/frameworks/kcoreaddons/-/merge_requests/166).
Because of this we define the macro ourselves, if we don't have it from kcoreaddons.

Once we depend on KF 5.88, we can drop the desktop files entirely.