Update vat account assignment during import
For imported transactions the automatic category assignment can be based
on historical transactions. In case none is found with an identical
amount, the one with the smallest difference in amount is used as
template to assign the categories to the imported transaction.

This transaction could lead to a category for which the vat account
assignment was changed in the meantime. Simply copying the splits of the
old transaction leads to problems in this case, as a) the wrong tax
category is used and b) the amounts are not updated correctly.

This change checks for such a modification of the tax category and
handles the import appropriately by assigning the current tax category
and updates all values correctly.