Don't use invalid price pairs
For some reason price entries of the form "from currency to security"
are contained in the database (xml and sql). Such entries don't disturb
the price retrieval of KMyMoney as it can swap from and to if needed,
but they cause the online price update to be confused and failing.

The reason why these entries appear is yet unclear and needs to be
investigated further. To avoid these entries to cause problems during
online price update, they will not be listed in the price updater. This
change also removes those entries when writing the data file (or the
database for that matter) the next time. A respective debug message is
printed to the console.

This does not finally fix the problem reported but it cures the
symptoms.

CCBUG: 450593
FIXED-IN: 5.1.3