Fix initial value loaded into transaction editor's category selector
Before this change was applied, when a transaction had no category
assigned, the first entry of the model was presented in the combo text
field even though it could not be selected. This change takes care, that
in case of a missing second split the text area remains cleared upon
loading of the transaction.

This amends commit a4ae60a4b0b484c3f680aba37b449e08918ad3cb