Don't treat post dates in the future as statement date
In case an imported statement does not provide a statement date it is
derived by using the newest post date of the transactions. It should not
be used though if that is in the future.

BUG: 456086
FIXED-IN 5.1.3

(cherry picked from commit 47d100b4cd841b999539db1d78ef9a3afe1141fa)