Update stock split if switching activity from/to dividend
The amount of shares of a dividend transaction could erroneously contain
a share amount different from zero which affects the amount of shares
held in that security account even though none were added/removed.

This change makes sure, that the amount of shares, the value and the
price are reset when the activity is switched to a dividend transaction
and reloads any previous values contained in the edit fields when
switching away from it.