Make account/category selection permanent
In some cases when an account or category was selected in the opened
popup treeview the selected account name was displayed in the widget but
the change was not persistent. This happened to me most often when I
pressed OK right after the selection of the account with the mouse was
performed.

This change fixes the problem.