Improve OFX importer
Allow fractions of security smaller than 1/100. A new security is
created with a default fraction of 1/100. In case the import contains a
transaction with more fractional digits, the fraction of the security is
automatically adjusted so that no information is lost.

Securities are also detected by their unique id not only by their name.
This allows to change the name after the import without re-creating the
security during the next import.

Security accounts are using the oldest postdate found in the OFX import
file as their opening date and not the current date. This prevents the
consistency check to report and fix the opening date later on.

BUG: 458447