Propagate leftover/overdrawn budget of one period into next
This adds a feature to reporting which propagates the budget difference of e.g. one month into the next.

![image](/uploads/349f41da7a2fef4e11094ab23d69eaa0/image.png)

Having an expense budget with 120 € each month and the following expenses (Jan 20 €, Feb 230 €, Mar 10 €) with the current report implementation one does not see how the overall budget is doing at the beginning of April as it only shows 120 €.

![image](/uploads/28e8fff04bdc52506013092d242caeda/image.png)

When enabling the new option, the picture becomes clearer:

![image](/uploads/ca69d34244b8909f3b4bedf9b77909cc/image.png)