Skip state reconciled when the transaction will vanish
When the "Hide reconciled transactions" filter is active and one clicks
on the C column the transaction disappears when it changes state to
reconciled. This change does not switch to reconciled in case the filter
is active and so the transaction stays where it is. One can still
reconcile the account or use the transaction editor to change the state
to reconciled.