Improve handling of reconciliation
The previous implementation caused all transactions marked as reconciled
to be hidden no matter if they are within the period of of the statement
or not. Also, transactions not marked reconciled but have a post date
prior to the start of the statement were not correctly displayed during
the reconciliation process. This caused the balances to become
"incorrect".

This has been changed to show all transactions from the date of the
first non-reconciled one even if that is prior to the last
reconciliation. All reconciled transactions that have a post date that
is within the period of the statement are also not hidden anymore and
taken into account when calculating the balances. Transactions younger
than the statement date are not taken into account for balance
calculation as before.

I am not sure if this fixes the reported problem 431373 but it is
certainly an improvement.

CCBUG: 431373
BUG: 438936
FIXED-IN: master