Don't show positive total variation forecast values in red
A total variation item could temporarily be negative during the
calculation of the forecast but the final result is positive. In this
case, when it turned negative the color was changed to the negative
scheme (red) but never got turned back to the default when it became
positive again.

(cherry picked from commit c8a4c2539adb8e2189c2e9449a0cc1f103b13cc1)