Completely update reference to account during import
When data of an account is imported from a file, the KMyMoney account
will get a reference to the imported account using the account number
found. This allows to pre-select the account during import.

In case the user decided to override that pre-selected account, the
reference on the old account was not removed and the data of the
selected account was not used in the process of the import.

This change removes the reference on the previous account and uses the
the selected account to import the data.

BUG: 466809
FIXED-IN: 5.1.4

(cherry picked from commit c6357119672f3b124ffcc18f5078b2d65c9f9153)