Improve consistency check
The consistency check complained about a missing price information for
the investment account opening date. A price may not be needed for that
date if the account is only referenced by a transaction at a later point
in time. The consistency check has been changed so that it only
complains when no price information is available for the date of first
usage.