Select date section only when receiving focus
Within the transaction editor, the initial section of the date was
selected upon start of the editor even if the date widget did not have
focus. This could mislead the user. This change takes care that the
selection of the date section only happens when the date widget
receives/has focus.