Maintain selection of multiple transactions
When the display of multiple selected transactions was interrupted by
non transaction items (e.g. online balance, date or reconciliation
markers) then the selection was not maintained when an operation was
performed on all transactions (e.g. mark as cleared). This change makes
sure that the selection is maintained after the operation has finished.