Show transaction info
Add popup to show all details of the selected (in the ledger) transaction(s), including KVP for the transaction and any splits.  There has been a suggestion to also allow this to anonymize the data and save to a file, to assist with reporting bugs.  While a good idea, I will leave it for a separate effort.