Allow to fix amount for foreign currency category in schedules
See https://discuss.kde.org/t/scheduled-transaction/3210 for use case

This option is enabled for a schedule if the selected category and
account use different currencies. For others it is disabled. It is also
disabled in case of split transactions.

When the scheduled transaction is entered into the ledger the user can
modify the amount. If the option is active, the amount for the (foreign)
category is not changed but the exchange rate of the transaction (not in
the price history) is adjusted instead.