Fix LibOFX detection
LibOFX uses two different version numbers which are not directly related
to each other. One is used for the package and the other as version for
the shared object library. CMake's find_package looks for the library's
so version whereas pkg_check_modules retrieves the package version.

This change makes sure to distinguish between the two numbers.