Adding payee-based links in the memo field of transactions
For this purpose, a new tab entry “Link Matching” has been
added to the processing interface for payees, which contains
two input fields, one for a pattern for identifiers and one
for a url template with a placeholder “%1”.

If these fields are filled for a recipient, a matching
identifier is searched for in the memo field in the general
ledger view for transactions that are assigned to this
recipient and, if found, marked accordingly and inserted
into the url template.

The link created in this way can be opened via the new
“Open URL” entry in the context menu, via a keyboard
shortcut if configured in the settings and in the form
view and editor of a transaction via a displayed link.

Simple examples of a pattern are \d{10} for a 10-digit
number or “\d+-\d+” for two blocks of numbers connected
with a hyphen.

BUG:437481
FIXED-IN:5.2.0