Assign tag to the category split
The initial implementation of tags assigned the selected tag to the
split for the account the transaction was entered. This is not useful as
the tag needs to be assigned to the category split.

This change makes sure that the tag is assigned to the category split.
Existing simple transactions can be fixed by opening them in the
transaction editor and saving them again. Multi split transactions need
to be fixed using the split editor.

The above described fix only works for individually selected
transactions. It is not supported when multiple transactions are
selected.

CCBUG: 480360
CCBUG: 406403