Fix reporting of tagged splits
Reporting tagged transactions did not work properly due to the
assignment of the tag to the wrong split. See commit 96a4ae27bb for a
fix to that problem.

For it to produce correct results in the report, the transactions may
have to be edited so that the tag is assigned to the correct split.

BUG: 406403
FIXED-IN: 5.2