Allow semi-colons in name matching imported from online sources
The semi-colon was used in version 4.8 and prior as a key separator.
Since 5.0 it has been replaced with \n but the old method is still
supported when the new one is not detected. Having a payee that was
imported from an online source and contained a semi-colon confused the
logic.

This change makes sure, that such import is handled properly by forcing
the string to always contain a \n even if there is only one key in the
list.