Keep account selection in case it is not changed
Since the QComboBox popup view in general only supports simple lists and
is based on the selection of rows, it did not pick up the row as part of
an item the subtree but used the item on the top level of the model with
the same row instead, but only if there was no change in selection.

This change takes care of supporting a tree view as popup with the
underlying account model.

BUG: 471210
FIXED-IN: 5.2