Show online price entries no matter if used or not
Online price entries for update were only shown if they are actually
used by any of the accounts. That lead to the fact that existing price
entries in the price table were not shown for online update if no
account was using them.

This change forces all price entries to be shown for update.

BUG: 485767
FIXED-IN: 5.2