Filter out invalid XML characters
Earlier versions of KMyMoney could leave invalid XML characters in the
data file (mostly in data imported from institutions). This causes the
XML reader to fail and the remainder of the file is lost.

This change introduces a filter that removes those invalid characters
before the file content is processed by the XML stream reader.