Prevent crash when widget's slot is called after destruction
A crash could happen when both amount widgets were modified (moved the
amount from one to the other widget), the focus remained in the
receiving widget and the enclosing dialog was closed. The focusOut event
handling caused the emission of the amountChanged signal as part of
the destruction of the amount widget that did not have focus anymore at
a time, when the enclosing CreditDebitEdit widget was already destroyed.

This change makes sure, that such signal/slot calls cannot happen by
explicitly disconnecting the connections when the enclosing
CreditDebitEdit is destroyed.