Fix function of return key on activity widget
When the return key was pressed when the activity widget had focus
inside the investment transaction editor, the event was discarded by the
widget stack. This is not the case for all other widgets of the editor.

This change makes sure, that the functionality is available even on the
activity widget.