Show liabilities/equities as positive numbers
A liability is usually a negative amount in terms of networth. Showing
the liabilities with a negative sign is confusing for users, though.

This change takes care of this and displays liabilites that are still
due as positive numbers and overpaid liablities as negative on the
homepage's asset and liability section, the account balance chart and
the ledger balance.

BUG: 486876
FIXED-IN: 5.2