Provide information about historic date range used for forecast
In some cases it might be unclear to the user why the budget information
in the forecast does not show the expected result. Since this is based
on the transactions found in a specific date range, displaying this date
range may help the user to understand what is going on.

CCBUG: 488231