Improve crypto strength of recover key by using a new 4096 bit key
The recover key used in the past does not provide the required crypto
strength anymore. 1024 bit keysize are considered to be unsafe. A new
recover key has been created and is used in version > 5.1.3 from now on.

A bit of cleanup also happened since the key information was held in
different places of the source code. This has now been unified and is
maintained in a single location so that the next switch to a new recover
key is lot quicker.

The key selection dialogs now don't show duplicates anymore to improve
the key selection process.