Split dialog: show amounts in transaction currency
When split was assigned to an account denominated in a different
currency than the transaction, the amount was shown in the account's
currency. This did not allow to sum up the amounts in the split dialog.

The amount is now shown in the transaction currency and hovering over
the amount shows the foreign amount in a tooltip.