Improve detection of currency formatting
QLocale::toCurrencyString() returns different strings on Windows and
Linux when using a value without fraction (e.g. 100). On Linux the
output in locale de_DE was "100 €" and the same code on Windows returned
"100,00 €". That caused the existing implementation of the format
detection to fail.

Changing the test value to 123.45 and enhancing the regex to extract the
parts solves the issue.

BUG: 491828
FIXED-IN: 5.2