Make sure that shares and value of a split have the same sign
For some (yet unknown) reason, the CSV importer creates splits that have
positive shares and a negative value. That is incorrect and needs to be
fixed.

This change makes sure, that no such split can enter the journal and
uses the sign of the shares in favor in case they differ.