Show warning of missing category assignment or category combo
A warning was shown when the amount of all splits of a transaction was
not balanced (i.e. not equal to zero) or a catgegory assignment was
missing while an amount was entered.

It makes more sense to show the warning frame and hint on the category
combo box in the latter case.