Add support for testing report generating
The support is based on a command line tool that loads
kmymoney files and exports all user-defined reports to
files that are then compared with reference files using
a cmake script. Any differences that occur are displayed
in the log and marked as errors. For each test file added
to the cmake build system a dedicated cmake test is added.

BUG:492891
FIXED-IN:5.2