Sync the investment view to the ledger view on first access
When an investment account is selected after startup of the application
in the ledger view and then the investment view is opened, a different
account may be selected in the investment view. This only happens the
first time after application start.

This change makes sure that even on the first time the views are synced
and the correct account is selected in the investment view.