Calculate loan values before showing on home page
When showing the receiving split of a loan schedule, the amout shown is
incorrect (huge negative value). This special value indicates, that the
actual values for the loan transaction need to be calculated based on
the loan terms (balance, interest rate, etc.). This calculation was
missing when processing the entry.

This change adds the necessary step and calculates the correct values
for the transaction to be displayed.