Fix conversion to base currency for reports containing categories
The previously used logic to recognize the conversion to the base
currency was not correct for the report mentioned in this error.

It was added to fix bug 448788 and worked there because the base
currency was taken from the first split, which happened to be the
same as the one from the file.

In general, when converting to the base currency, the base currency
of the file should be used and not the currency from the first split,
which has been corrected with this commit.

BUG:382245
CCBUG:448788
FIXED-IN:5.2.0