Correction of expected test report values for “test-423870.xml”
After the problem reported in error 382245 has been fixed, the
expected values in reports of the mentioned test have changed
and need to be adjusted.

As the target accounts are managed in the US currency, the
corresponding splits must also be splits must also be issued
in this currency, as mentioned in the ticket.

CCBUG:423870