Restore indentation for reports based on pivot tables
The commit below increased the indentation to 16 times the
value previously used, which significantly increased the
width of the report when exporting to an HTML file; this
has now been corrected.

If a higher value is required, this should be done via an
option in the settings.

Fixup for commit f0a9807dd.

CCBUG:492831