cmake: Add support to generate the doc as normal html page
This support is primarily intended for developing documentation
to avoid the more complex detour via meinproc and khelpcenter
as viewer. However, it can also be used to provide offline
documentation bundled to the application. The only dependencies
required are the xsl stylesheets for docbook, which are already
a dependency of kdoctools.

The cmake macro add_docbook_manual() was created with the
support of chatgpt, which helped to reduce the required
development time.