cmake: Add support to generate the doc as normal html page
This support is primarily intended for developing documentation to
avoid the more complex detour via meinproc and khelpcenter as viewer.
However, it can also be used to provide offline documentation. The
only dependencies required are the xsl stylesheets for docbook,
which are already a dependency of kdoctools.

The cmake macro add_docbook_manual() was created with the support
of chatgpt, which helped to reduce the required development time.

The cmake macros add_auto_option() and check_auto_option() originate
from the D-Bus project, which was originally developed by Ralf Habacker.