Don't change the payee during autofill
When a payee is selected using the keyboard navigation and pressing the
return key, opening the autofill dialog caused the payee to be changed
when the autofill operation was completed. This change postpones the
autofill operation until the event loop is reached after the payee has
been selected with the return key.

BUG: 497120
FIXED-IN: 5.2