Improve handling of font changes in ledger
When the user changed the font settings for the ledger, some widgets do
not receive the updated font. These are the amount edit widgets in the
transaction editors and the split view. This change updates takes care
of updating all those widgets with the new font when it is changed in
the settings dialog.

BUG: 497255
FIXED-IN: 5.2