Include paid out dividends
When an investment report was configured to include only the owned
shares, the paid out dividends where not shown and part of the
calculation of returns. This change displays them in the report and
includes them in the calculation of returns.

See https://bugs.kde.org/show_bug.cgi?id=452692