Fix fractional part in balance column of ledger views
For currencies the fractional part was always 2 digits no matter what
the fraction setting of the account was set to. Since the balance column
could either show currencies or securities, both models have to be
searched for the account's currencyId. This change takes care of it. In
case the currency is not found, the id is searched in the securities. If
it is still not found it falls back to the default of 2 digits.