Fix calculation of starting balance during reconciliation
When reconciling an account, the starting balance is calculated by
subtracting the amount of all transactions with a posting date later
than the date of the last reconciliation. Unfortunately, transactions on
that very day were also subtracted which lead to a wrong starting
balance. This change fixes the calculation to include them.

BUG: 498080
FIXED-IN: 5.2