Extend adding debug information for generating xml export
The ListTable class receives a new method addSourceLine() for simple
output of source text lines of the associated map keys in the XML
output as TableRow attribute “ctSourceLines”.
This makes it possible to recognize at which source text positions
outputs are defined, whether a key is output multiple times or
not at all.

Corresponding outputs are currently generated for the ctValue
and ctPrice keys.

CCBUG:498033