Keep track of tooltip in WidgetHintFrame
Currently, the WidgetHintFrame methods show() and hide() provide a
method to assign a tooltip to a widget. In case a tooltip is set
directly on the widget it will be lost and also, the application not
only has to maintain if the frame shall be shown or not but also keep
track of the tooltip.

This change keeps track of the tooltip directly set by the application
and the one that should be visible when the frame is shown and updates
the tooltip accordingly. This allows the an application to change the
tooltip even if the frame is shown.

For backward compatibility show() and hide() still allow to set the
tooltip text.