Fix MSVC link error caused by unresolved symbol
The failure looks as:
  kmymoneyadaptor.cpp.obj : error LNK2019: unresolved external symbol
  QDBusAbstractAdaptor::qt_metacast

By calling the cmake macro qt_add_dbus_adaptor() the mentioned
C++ file is created and therefore probably needs the import
library of Qt6DBus which is not included in the link list
because KMM_DBUS is not set on Windows. Qt5 is probably always
linked with Qt5DBus, so that this problem does not occur there.