Fix matching of loan payment transactions
When a loan payment transaction is imported and matched against a
scheduled loan payment, the actual split between interest and principal
need to be calculated before the transaction can be entered into the
ledger. This was not done resulting in incorrect huge values to be shown
in the ledger after import. This change takes care of the calculation
before the matched transaction is entered.