Fix incorrect display in the balance history report
In this specific case, the values were between 120 and 240, which,
according to current knowledge, is not handled correctly by
kdiagram above a certain diagram width.

In this case, an additional vertical distance of 10 was added
in order to obtain an entry above the displayed values on the
vertical axis for smaller values.

CCBUG:499211
FIXED-IN:5.2