Fix inclusion of investment transactions in querytable reports
For an investment transaction, the query table report processes the
split that references the stock account as the first split. In cases
when an account filter was set, these transactions were not included in
the account even when their investment parent account was selected. A
direct selection of the stock account is not possible because the
selection widget does not show them.

This change fixes the problem by also checking for the parent account to
be selected or not in case the stock account is not part of the
selection.