Differentiate between edit and enter scheduled transaction
When the due date of a scheduled transaction falls on a banking holiday,
the due date in the schedule view is probably adjusted to the next or
previous processing day based on the selected option of the schedule.

When the schedule is about to be entered the option is ignored and the
original next due date is used to. The user would have to change the
date manually. This change uses the adjusted date when the transaction
editor is used during entry of the scheduled transaction to solve the
problem.

BUG: 499459
FIXED-IN: 5.2