Add gdb pretty printer for MyMoneyMoney class
With Windows/gdb, 'set auto-load safe-path /' must be used
before starting an executable file to activate the loading
of a pretty printer.

Gdb pretty printers work in kdevelop and the gdb window in
QtCreator with the 'p' command, but currently not in the
symbol display due to a bug in the QtCreator gdb bridge
(https://bugreports.qt.io/browse/QTCREATORBUG-32480).

BUG:499752
FIXED-IN:5.2