Preload account selector when importing CSV statements
When CSV statement files don't include any hint for which account the
data is included, a dialog is presented to the user to select the
account to which the data shall be imported.

This change presets this selector with the currently selected account.
In my use case, I select the account in the ledger, click on the icon to
access the website and then use web-connect to download the CSV file
directly into KMyMoney. With this new functionality, the account is
already preselected and there's no need to pick it from the list
anymore.

The selection is also made available for the OFX importer but its usage
is not implemented yet.