Apply file fixes also to exported reports
With this commit applying file fixes has been moved
to the MyMoneyFile class to let the unit tests for
reports show the same results as when loading a
kmymoney file by the ui.

With this shift, notes issued via the UI are now
output as debug output on the console, as MyMoneyFile
does not support UI interaction.