Add column for return value in investment reports
Outputting the return as a value increases the information
content of the report.

With the addition of this column, the order of the columns displaying
returns in percent has been changed to obtain a more natural order.
After the return in a currency comes the return in percent, followed
by the annual return and the extended internal rate of return.

CCBUG:396301