Make sure a webconnect server is running at any time
When multiple instances of KMyMoney are started the first
one becomes the webconnect server and all others the
webconnect clients. If the frist instance is quit, the webconnect
server stopped working and when a webconnect import was
started, this failed and KMyMoney used to open the last opened
file which is still open in the second instance.

This change makes sure that a webconnect server is running
at any time to resolve the issue.