Add option to specify the extension for backups
Apparently, Microsoft Onedrive cannot handle the typical
Unix style backup identifier ~ (tilde) used by the automatically
created backup files when the data is saved. This option allows
to overwrite this character which still remains the default.
The field in the configuration dialog cannot be left empty. If
cleared, the default is loaded.