Improve import summary dialog
Importing transactions either from a file from disk, via web-connect or
direct connect to the bank shows a summary at the end. This summary has
been improved to present the data in a treeview per institution and for
each account.

When no transaction was added because of duplicate detection the entries
will be grayed out so that the focus is directed on new transactions.