Fix widget handling in transaction and split editor
Added an invisible column to the model which if not present
has the side effect when the left most column of the journal
or split model is moved right that the QTableView paints a
grid line in the area that is covered by the editor widget.

Synced the painting logic in journal and split delegate.

When first used, this will reset some settings of the
views to their defaults which was needed due to the
new column.

BUG: 503095
FIXED-IN: 5.2