Improve widget to open ledger for account
The tab marked with a plus sign in the ledger view allows
to open the ledger of another account. The size of this
widget changed erratically when text was typed which
caused the drop down view also shrink so that the names
of accounts could not really be determined. This change
fixes the problem.