Force the price column to be added only if prices are available
In cases where no prices are available, in reports it is ensured
that the price used for currency conversion is always 1.0, which
makes it unnecessary to display the price column.

When loading files, it is ensured that the price column is not
displayed in this case in all reports.

BUG:502604
FIXED-IN:5.2