Don't use line breaks in tabbed view
The title of the scheduled transaction view is shown in two lines
separated with a line break. This does not look nice when the views are
shown in tabbed view. This change replaces the line break with a space
and vice versa when switching the view type.

(cherry picked from commit be35cd5bb2a154ee073e3105221c25cd9cd10f4e)