Fix deleting multiple splits
In case more than one split was deleted in the split editor they were
not all removed when the transaction was actually saved. This was caused
by iterating over a list that is modified in the loop. Iterating over a
copy of the list to be modified resolves the problem.