Fix deleting multiple splits
In case more than one split was deleted in the split editor they were
not all removed when the transaction was actually saved. This was caused
by iterating over a list that is modified in the loop. Iterating over a
copy of the list to be modified resolves the problem.

(cherry picked from commit 84cbf822dfd32fa413014800b444ca31c6e4ad46)