Fixup imported transactions that contain confusing values
Meryll Lynch exports some strange transactions, where they create a sell
transaction with an amount not equal to zero and a price being zero and
also shares being zero.

Such transaction causes a division by zero later on at some point, when
the value is divided by the number of shares to calculate the price. We
have no idea why Merrill Lynch calls this a sell, as it appears to just
be some sort of dividend or other payout of the amount. We simply
convert this into an interest transaction.

(cherry picked from commit cb810055443cdb5a8dabeb3f87da36b826a8080f)