Prevent changing the reconciliation status
In case one or more transaction are selected and the user clicks on the
C column in a grouping header or non-selected transaction, the
reconciliation status of the selected transactions changes.

This change resolves the problem by limiting the click functionality to
a selected row.

(cherry picked from commit be64db3dcd8336e4a39ce3cd94542d245509ff22)