Fix infinite loop when open existing database created with 5.1.3
The configuration of reports needed to be modified in certain
cases. This is done automatically when loading an older
data file, but the implementation was missing for the
database backend. This causes an infinite loop when
opening the database and renders the application useless.

This change provides the necessary implementation.

BUG: 506569
FIXED-IN: 5.2.1
(cherry picked from commit fe163b37b2e3ebb599b74bca21a113dc8135d624)