Allow more data to be entered when searching for old txn
The search for old transactions for a payee was only allowed when no
amount, no memo and no category was entered. This is a bit tight and
has been changed to only be disabled when a category is entered before
the payee.

BUG: 506625
FIXED-IN: 5.2.1

(cherry picked from commit 8c019fb0d9ad9f47819a472f5bcfb7f10cc74931)