Fix problems with invalid reinvestment transactions in constency check
Some institutions apparently provide a negative amount for a reinvested
dividend while the number of shares is positive. This was somewhat
ignored by versions prior to 5.2.0 but will now be caught by the
consistency check.

This change provides a fix to the consistency check so that the affected
invalid transactions are corrected.

BUG: 506949
FIXED-IN: 5.2.1
(cherry picked from commit 1f6662e07a41d9050e3fec89e35f88d23db4be6b)