Filter on account names in all splits of a transaction
The filter only searched in the account name of the split shown in the
ledger. This change makes sure that the entry matches if the provided
search text is found in any account name of all splits of the
transaction.

BUG: 507104
FIXED-IN: 5.2.1
(cherry picked from commit 80ad52b03009a31a877bb8726e4a42880b062aa3)