Remove only characters selected by user
When modifying the text in the amount edit widget by using the backspace
or delete key it could happen that more than one character was removed.
This confuses the user, epsecially if the decimal symbol gets removed as
part of this.

This change ensures that only the characters selected by the user are
removed.

BUG: 507195
FIXED-IN: 5.2.1