Use complete width for account info on home page
When the account limit information was not selected to be displayed, a
gap appears to the right of the balance column. This change makes sure
to use the full width even if the limits are not displayed.

BUG: 507098
FIXED-IN: 5.2.1
(cherry picked from commit 90dd25bdb92d34631e78c1e0c743cec81c536e3b)