Improve keyboard handling in date edit widget
Pressing Ctrl+A selects all characters of the date. If only a part of
the date is selected but the text to be pasted is larger, then the
widget is cleared first. Selection using Ctrl and Shift works as
expected and does not switch sections of the date anymore.

BUG: 507244
FIXED-IN: 5.2.1