Update label when security changes
When the security widget was left with the TAB key the labels for
certain widgets were not updated according to the settings for the
selected security. This change fixes the problem.

BUG: 507664
FIXED-IN: 5.2.2
(cherry picked from commit 9f1c7dd0c1c76bdaa4b98b267c787613e94a1e64)

Was fixed in master for 5.2.1 but not carried over to the stable branch.